rehab_system/requirements.txt
# 3D Hand Rehabilitation System
# Professional biomedical application for hand physiotherapy

PyQt5==5.15.9
vtk==9.2.6
numpy==1.24.3
scipy==1.11.1
pandas==2.0.3
matplotlib==3.7.2
websocket-client==1.6.1
pyserial==3.5
